requirements.txt
# AWS SDK and CDK
boto3==1.34.34
aws-cdk-lib==2.118.0
constructs>=10.0.0,<11.0.0

# AWS Lambda Powertools
aws-lambda-powertools==2.31.0

# Testing
pytest==7.4.4
pytest-cov==4.1.0
pytest-mock==3.12.0
hypothesis==6.98.3
moto==5.0.0

# Utilities
python-dotenv==1.0.0
pydantic==2.5.3
